Checkpoint (Karthik): snapshot of all progress till Nov 8, 2025

+++ Controllers/CoursesController.cs
@@ -1,12 +1,10 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using EduvisionMvc.Data;
 using EduvisionMvc.Models;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace EduvisionMvc.Controllers
 {
@@ -19,26 +17,32 @@ namespace EduvisionMvc.Controllers
             _context = context;
         }
 
+        private void PopulateDepartmentsSelectList(object? selected = null)
+        {
+            ViewBag.Departments = new SelectList(
+                _context.Departments.OrderBy(d => d.Name),
+                "Id", "Name", selected);
+        }
+
         // GET: Courses
         public async Task<IActionResult> Index()
-        {
-            return View(await _context.Courses.ToListAsync());
-        }
+{
+    var courses = await _context.Courses
+        .Include(c => c.Department)   // ✅ include this
+        .ToListAsync();
+    return View(courses);
+}
+
 
         // GET: Courses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var course = await _context.Courses
+                .Include(c => c.Department)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (course == null)
-            {
-                return NotFound();
-            }
+            if (course == null) return NotFound();
 
             return View(course);
         }
@@ -46,15 +50,14 @@ namespace EduvisionMvc.Controllers
         // GET: Courses/Create
         public IActionResult Create()
         {
+            PopulateDepartmentsSelectList();
             return View();
         }
 
         // POST: Courses/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,Code,Title,Credits,Dept")] Course course)
+        public async Task<IActionResult> Create([Bind("Id,Code,Title,Credits,DepartmentId")] Course course)
         {
             if (ModelState.IsValid)
             {
@@ -62,36 +65,28 @@ namespace EduvisionMvc.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+            PopulateDepartmentsSelectList(course.DepartmentId);
             return View(course);
         }
 
         // GET: Courses/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var course = await _context.Courses.FindAsync(id);
-            if (course == null)
-            {
-                return NotFound();
-            }
+            if (course == null) return NotFound();
+
+            PopulateDepartmentsSelectList(course.DepartmentId);
             return View(course);
         }
 
         // POST: Courses/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Code,Title,Credits,Dept")] Course course)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,Code,Title,Credits,DepartmentId")] Course course)
         {
-            if (id != course.Id)
-            {
-                return NotFound();
-            }
+            if (id != course.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -99,37 +94,28 @@ namespace EduvisionMvc.Controllers
                 {
                     _context.Update(course);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!CourseExists(course.Id))
-                    {
+                    if (!_context.Courses.Any(e => e.Id == course.Id))
                         return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    else throw;
                 }
-                return RedirectToAction(nameof(Index));
             }
+            PopulateDepartmentsSelectList(course.DepartmentId);
             return View(course);
         }
 
         // GET: Courses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var course = await _context.Courses
+                .Include(c => c.Department)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (course == null)
-            {
-                return NotFound();
-            }
+            if (course == null) return NotFound();
 
             return View(course);
         }
@@ -140,18 +126,9 @@ namespace EduvisionMvc.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var course = await _context.Courses.FindAsync(id);
-            if (course != null)
-            {
-                _context.Courses.Remove(course);
-            }
-
+            if (course != null) _context.Courses.Remove(course);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool CourseExists(int id)
-        {
-            return _context.Courses.Any(e => e.Id == id);
         }
     }
 }
