+++ Controllers/InstructorsController.cs
@@ -1,12 +1,10 @@
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using EduvisionMvc.Data;
 using EduvisionMvc.Models;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace EduvisionMvc.Controllers
 {
@@ -19,26 +17,31 @@ namespace EduvisionMvc.Controllers
             _context = context;
         }
 
+        private void PopulateDepartmentsSelectList(object? selected = null)
+        {
+            ViewBag.Departments = new SelectList(
+                _context.Departments.OrderBy(d => d.Name),
+                "Id", "Name", selected);
+        }
+
         // GET: Instructors
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Instructors.ToListAsync());
+            var q = _context.Instructors
+                .Include(i => i.Department)
+                .AsNoTracking();
+            return View(await q.ToListAsync());
         }
 
         // GET: Instructors/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var instructor = await _context.Instructors
+                .Include(i => i.Department)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (instructor == null)
-            {
-                return NotFound();
-            }
+            if (instructor == null) return NotFound();
 
             return View(instructor);
         }
@@ -46,15 +49,14 @@ namespace EduvisionMvc.Controllers
         // GET: Instructors/Create
         public IActionResult Create()
         {
+            PopulateDepartmentsSelectList();
             return View();
         }
 
         // POST: Instructors/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Id,FirstName,LastName,Email")] Instructor instructor)
+        public async Task<IActionResult> Create([Bind("Id,FirstName,LastName,Email,DepartmentId")] Instructor instructor)
         {
             if (ModelState.IsValid)
             {
@@ -62,36 +64,27 @@ namespace EduvisionMvc.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+            PopulateDepartmentsSelectList(instructor.DepartmentId);
             return View(instructor);
         }
 
         // GET: Instructors/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var instructor = await _context.Instructors.FindAsync(id);
-            if (instructor == null)
-            {
-                return NotFound();
-            }
+            if (instructor == null) return NotFound();
+
+            PopulateDepartmentsSelectList(instructor.DepartmentId);
             return View(instructor);
         }
 
         // POST: Instructors/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,LastName,Email")] Instructor instructor)
+        public async Task<IActionResult> Edit(int id, [Bind("Id,FirstName,LastName,Email,DepartmentId")] Instructor instructor)
         {
-            if (id != instructor.Id)
-            {
-                return NotFound();
-            }
+            if (id != instructor.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -99,37 +92,28 @@ namespace EduvisionMvc.Controllers
                 {
                     _context.Update(instructor);
                     await _context.SaveChangesAsync();
+                    return RedirectToAction(nameof(Index));
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!InstructorExists(instructor.Id))
-                    {
+                    if (!_context.Instructors.Any(e => e.Id == instructor.Id))
                         return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    else throw;
                 }
-                return RedirectToAction(nameof(Index));
             }
+            PopulateDepartmentsSelectList(instructor.DepartmentId);
             return View(instructor);
         }
 
         // GET: Instructors/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (id == null) return NotFound();
 
             var instructor = await _context.Instructors
+                .Include(i => i.Department)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (instructor == null)
-            {
-                return NotFound();
-            }
+            if (instructor == null) return NotFound();
 
             return View(instructor);
         }
@@ -140,18 +124,9 @@ namespace EduvisionMvc.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var instructor = await _context.Instructors.FindAsync(id);
-            if (instructor != null)
-            {
-                _context.Instructors.Remove(instructor);
-            }
-
+            if (instructor != null) _context.Instructors.Remove(instructor);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool InstructorExists(int id)
-        {
-            return _context.Instructors.Any(e => e.Id == id);
         }
     }
 }
